requirements.txt
asgiref==3.11.1
Django==5.2.11
django-ckeditor==6.7.3
django-ckeditor-5==0.2.19
django-js-asset==3.1.2
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
